GH-935

+++ MultisigWallet/MultisigWalletDomainModel/NotificationService/Messaging/CommunicationDomainService.swift
@@ -11,7 +11,7 @@ public class CommunicationDomainService {
     public func deletePair(walletID: WalletID, other address: String) throws {
         let wallet = DomainRegistry.walletRepository.find(id: walletID)!
         let deviceOwnerAddress = wallet.owner(role: .thisDevice)!.address
-        let eoa = DomainRegistry.externallyOwnedAccountRepository.find(by: deviceOwnerAddress)!
+        guard let eoa = DomainRegistry.externallyOwnedAccountRepository.find(by: deviceOwnerAddress) else { return }
         let signature = DomainRegistry.encryptionService.sign(message: "GNO" + address, privateKey: eoa.privateKey)
         let request = DeletePairRequest(device: address, signature: signature)
         try DomainRegistry.notificationService.deletePair(request: request)
